Refactoring: project structure update
