+++ app/src/main/java/com/dyshuk/android/upworkredesigncompose/ui/components/JobListItem.kt
@@ -1,0 +1,2 @@
+package com.dyshuk.android.upworkredesigncompose.ui.components
+
